Finishing split into library and executable target
